% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_set.R
\name{are_in.mesh}
\alias{are_in.mesh}
\title{Checks if some points belong to a given mesh}
\usage{
are_in.mesh(X, mesh)
}
\arguments{
\item{X}{points to check}

\item{mesh}{mesh identifying the set which X may belong}
}
\description{
Checks if some points belong to a given mesh
}
\examples{
X = matrix(runif(100),ncol=2);
inside = are_in.mesh(X,mesh=geometry::delaunayn(matrix(c(0,0,1,1,0,0),ncol=2),output.options =TRUE))
print(inside)
plot(X,col=rgb(1-inside,0,0+inside))
}
