% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{branin}
\alias{branin}
\title{This is a simple copy of the Branin-Hoo 2-dimensional test function, as provided in DiceKriging package.
The Branin-Hoo function is defined here over [0,1] x [0,1], instead of [-5,0] x [10,15] as usual.
It has 3 global minima : x1 = c(0.9616520, 0.15); x2 = c(0.1238946, 0.8166644); x3 = c(0.5427730, 0.15)}
\usage{
branin(x)
}
\arguments{
\item{x}{a 2-dimensional vector specifying the location where the function is to be evaluated.}
}
\value{
A real number equal to the Branin-Hoo function values at x
}
\description{
This is a simple copy of the Branin-Hoo 2-dimensional test function, as provided in DiceKriging package.
The Branin-Hoo function is defined here over [0,1] x [0,1], instead of [-5,0] x [10,15] as usual.
It has 3 global minima : x1 = c(0.9616520, 0.15); x2 = c(0.1238946, 0.8166644); x3 = c(0.5427730, 0.15)
}
