% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_set.R
\name{combn.design}
\alias{combn.design}
\title{Generalize expand.grid() for multi-columns data. Build all combinations of lines from X1 and X2. Each line may hold multiple columns.}
\usage{
combn.design(X1, X2)
}
\arguments{
\item{X1}{variable values, possibly with many columns}

\item{X2}{variable values, possibly with many columns
combn.design(matrix(c(10,20),ncol=1),matrix(c(1,2,3,4,5,6),ncol=2))
combn.design(matrix(c(10,20,30,40),ncol=2),matrix(c(1,2,3,4,5,6),ncol=2))}
}
\description{
Generalize expand.grid() for multi-columns data. Build all combinations of lines from X1 and X2. Each line may hold multiple columns.
}
