% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{EvalInterval.function}
\alias{EvalInterval.function}
\title{eval function and cast result to a list of y, y_low, y_up (possibly NA)}
\usage{
EvalInterval.function(fun, X, vectorized = FALSE, dim = ncol(X))
}
\arguments{
\item{fun}{function to evaluate}

\item{X}{matrix of input values for fun}

\item{vectorized}{whether fun is vectorized or not}

\item{dim}{dimension of input values for fun if}
}
\value{
list of y, y_low, y_up
}
\description{
eval function and cast result to a list of y, y_low, y_up (possibly NA)
}
