% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{fade}
\alias{fade}
\title{be careful that fade(.., alpha=0) means total fading, while fade(.., alpha=1) means no fading}
\usage{
fade(
  color = "red",
  alpha = seq(from = 0, to = 1, length.out = 5),
  plot = FALSE
)
}
\description{
be careful that fade(.., alpha=0) means total fading, while fade(.., alpha=1) means no fading
}
