% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_set.R
\name{is_in.p}
\alias{is_in.p}
\title{Test if points are in a hull}
\usage{
is_in.p(x, p, h = NULL)
}
\arguments{
\item{x}{points to test}

\item{p}{points defining the hull}

\item{h}{hull itself (built from p if given as NULL (default))}
}
\description{
Test if points are in a hull
}
\examples{
is_in.p(x=-0.5,p=matrix(c(0,1),ncol=1))
is_in.p(x=0.5,p=matrix(c(0,1),ncol=1))
is_in.p(x=matrix(-.5,ncol=2,nrow=1),p=matrix(c(0,0,1,1,0,0),ncol=2))
is_in.p(x=matrix(.25,ncol=2,nrow=1),p=matrix(c(0,0,1,1,0,0),ncol=2))
is_in.p(x=matrix(-.5,ncol=3,nrow=1),p=matrix(c(0,0,0,1,0,0,0,1,0,0,0,1),ncol=3,byrow = TRUE))
is_in.p(x=matrix(.25,ncol=3,nrow=1),p=matrix(c(0,0,0,1,0,0,0,1,0,0,0,1),ncol=3,byrow = TRUE))
}
