% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_set.R
\name{min_dist}
\alias{min_dist}
\title{Minimal distance between one point to many points}
\usage{
min_dist(x, X, norm = rep(1, ncol(X)))
}
\arguments{
\item{x}{one point}

\item{X}{matrix of points (same number of columns than x)}

\item{norm}{normalization vecor of distance (same number of columns than x)}
}
\value{
minimal distance
}
\description{
Minimal distance between one point to many points
}
\examples{
min_dist(runif(3),matrix(runif(30),ncol=3))
}
