% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_set.R
\name{points_in.mesh}
\alias{points_in.mesh}
\title{Extract points of mesh which belong to the mesh triangulation (may not contain all points)}
\usage{
points_in.mesh(mesh)
}
\arguments{
\item{mesh}{mesh (list(p,tri,...) from geometry)}
}
\value{
points coordinates inside the mesh triangulation
}
\description{
Extract points of mesh which belong to the mesh triangulation (may not contain all points)
}
