% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_set.R
\name{points_out.mesh}
\alias{points_out.mesh}
\title{Extract points of mesh which do not belong to the mesh triangulation (may not contain all points)}
\usage{
points_out.mesh(mesh)
}
\arguments{
\item{mesh}{(list(p,tri,...) from geometry)}
}
\value{
points coordinates outside the mesh triangulation
}
\description{
Extract points of mesh which do not belong to the mesh triangulation (may not contain all points)
}
