% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{expand.grids}
\alias{expand.grids}
\title{Create a Data Frame from all combinations of factor variables}
\usage{
expand.grids(d = length(list(...)), ...)
}
\arguments{
\item{d}{number of variables (taken in following arguments with modulo)}

\item{...}{variables to combine, as arrays of values}
}
\value{
data frame of all possible combinations of variables values
}
\description{
Generalization of base::expand.grid to more than 2 variables.
}
